[feature] Upgrade to "jQAssistant 1.6".
